Reorganize and complete test function
This fails currently for unknown reasons if an ROI is selected
